Support cross-running 32-bit efibootmgr build on 64-bit kernel
To comply UEFI spec, build from the existing implementation
can run on a kernel with the same architecture only, either
64-bit or 32-bit.

This patch provides support to run a 32-bit build on a 64-bit
kernel when the new compiling option is enabled.

When this option is not enabled, the build is compiled in the
same way of the existing. It is not effective either in 64-bit
build.

Note: We cannot test efibootmgr with the option enabled in adb
shell on a 64-bit kernel in Android build so far. There is no EFI
service exposed from BIOS in legacy mode.

Change-Id: Ie5f300d6061fc2cc2cd2053d266b02f672164e7c
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>